m5stickc-plus/collect-data/pyproject.toml
[project]
name = "collect-data"
version = "0.1.0"
description = ""
authors = [
    {name = "katatoshi",email = "15307563+katatoshi@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pyserial (>=3.5,<4.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
